Add file-handles migration mode
Add preserialization code to generate file handles, and allow
serializing them.

One important thing to note is that we want to drop the DAC_READ_SEARCH
capability if it is not needed.  However, the --migration-mode is
specified on the source, so we cannot use it as an indicator of whether
DAC_READ_SEARCH is needed or not.  Users therefore have to manually
indicate the need to retain this capability if migrating using file
handles by passing --modcaps=+dac_read_search to the destination
instance.  I found no better way than to just document this in both the
README file and our --help text.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>