Extract serialized::Inode.deserialize_root_node()
Put the bit of code the deserialization module uses to open the root
inode into a separate function.

For the file-handles migration mode, we are going to introduce a mount
FD map, i.e. a structure that maps the source’s mount IDs to locations
inside the shared directory, i.e. paths relative to the root node.  We
need to open all these locations so we can resolve the file handles,
which requires opening the root node before any other node can be
deserialized.  Having this separate function will make that more
sightly.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>