Move necessity check into update_IMI()
Whether it is necessary to update an inode’s migration info (IMI) when
the associated inode is renamed depends on whether we are currently
preparing for migration, and whether the IMI even contains a path – when
we introduce file-handle-based migration, we don’t want to discard
file-handled-based IMI just because a node is renamed.

Do both of those checks inside of `update_inode_migration_info()`, which
means removing the check from its caller `rename()`.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>