Invalidate inode’s paths when necessary
When an inode is overwritten by a different inode (e.g. through rename
or mkdir[1]), and we are preparing for migration and the overwritten
inode has path-based migration info on it, we should invalidate that
info so it isn’t sent to the destination.

This fixes cases where the guest replaces an inode (either via rename,
or unlink + creat) but keeps the previous inode open; on migration, we
would then transmit the outdated path for the inode that previously
existed there, making the destination open the new (different) inode,
leading potentially to data corruption.

With this patch, we will instead tell the destination that we don’t have
a valid path to the old inode, which will lead to errors (depending on
--migration-on-error), but at least not data corruption.

[1] Note that functions like mkdir, mknod, symlink, or link will
    generally return EEXIST when the destination already exists,
    refusing to overwrite it.  So in these cases, it isn’t strictly
    necessary to invalidate the already existing inode’s path, given it
    won’t be overwritten anyway.  Anyway, I feel like it’s cleaner not
    to rely on this behavior and just be consistent in all functions
    that might theoretically overwrite something.
    (Maybe also note that generally, that EEXIST is going to be
    generated in the guest already, so such operation shouldn’t even
    reach virtiofsd.)

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>