Upgrade vhost dependencies
Upgrade the vhost and vhost-user-backend crate dependencies to a version
that includes support for VHOST_USER_RESET_DEVICE.

This requires bumping the vm-memory and virtio-queue version
dependencies, too.

Support for `FS_*` requests has been removed from the vhost crate in its
commit 569ef077facafe21f16c0e65ff826b26d3c9a0b3 (“vhost_user: Remove
support for FS_* requests”), so we have to drop the fs_cache_req_handler
module that depends on and would handle those messages, and all code
relating to setupmapping/removemapping.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>