Double-check all paths in preserialization
After running a path-finding preserialization operation (i.e. finding
paths for all inodes in our inode store), we should double-check them:
When a path is invalidated by a guest (e.g. unlinked or overwritten by
something else), we should invalidate the respective inode’s migration
info.  (We don’t do that currently, but we will.)  It’s possible that
this happens concurrently to the preserialization operation setting that
exact inode’s migration info to point to the old path, which is why we
need to double-check all inodes’ migration info.

Reuse our --migration-confirm-paths code for this.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>