Add host/guest UID/GID translation code
The `IdMap` type allows translating between the host and guest domains
of UIDs and GIDs, using separate maps for either direction, and allowing
squashing (n:1 mappings) as desired.

To be able to map efficiently, we need a data structure that can map
ranges of values; as far as I know, the Rust standard library does not
include such a data structure, so this commit adds a new dependency: The
btree-range-map crate provides exactly the kind of structure we need
(`RangeMap`).  All dependencies pulled in this way (btree-range-map,
btree-slab, cc-traits, smallvec) are licensed Apache-2.0 OR MIT, like
most of our current dependencies.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>