Introduce explicit UID/GID types
We want to implement an internal mapping between guest and host UIDs and
GIDs.  To ensure this mapping is complete and we will not miss anything,
we need to be able to clearly separate between the two kinds.  The new
types added here (GuestUid, GuestGid, HostUid, HostGid) can do that.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>