PassthroughFsV2
Add a new version of our serialized format, `PassthroughFsV2`.  This
version includes a map of mount paths, i.e. a mapping of mount IDs to
paths in the shared directory.  Such a map is necessary to be able to
open file handles from the source on the destination, because the
source’s mount IDs need not be valid on the destination.

Adding new fields necessitates a version bump: We always want to be
backwards compatible, i.e. new versions should be able to parse old
formats.  However, adding `mount_paths` to `PassthroughFsV1` would be an
incompatible change: New versions would expect this field, old versions
would not provide it; deserializing data that does not include the field
will fail, and not e.g. deserialize an empty map.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>