Build mount FD map on deserialization
To use (open) a file handle, we need an open FD on the file handle’s
filesystem.  To allow this, every `OpenableFileHandle` object includes a
reference to such an FD, and one such FD exists for every mount ID for
which we have a file handle (filesystems are identified by their mount
ID).

For migration, we need to serialize these references somehow.  Of course
we cannot serialize those FDs as file handles because we need them to be
able to use file handles in the first place.

We also cannot just pass the mount ID and let the destination open a
respective FD by parsing /proc/self/mountinfo (which is how we normally
generate mount FDs) because mount IDs from the source are generally not
valid on the destination.

Therefore, the source needs to generate a map of (its) mount IDs to
paths inside of the shared directory that are on that mount ID.  The
destination then opens those paths, and thus has its mount FDs.  This
assumes that the destination uses the same mount structure as the
source, which I believe is reasonable for the file-handles migration
mode.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>