support idmapped mounts
Idmapped mounts (VFS idmap) is a Linux kernel feature, which allows
to create a special kind of a bind mount with an idmapping attached to it.
This makes a mount-level "shifted" vision of a filesystem objects
ownership. For more details please refer to [1], [2], [3], [4].

Most of dirty works is done on the kernel-side [1], but we also have to
do some changes on the virtiofsd side to take idmapping into account.

To support idmapped mounts from virtiofsd side, let's:
- enable FUSE_ALLOW_IDMAP extension to signal the guest kernel that virtiofsd
is aware of idmapped mounts feature
- ensure that virtiofsd does not perform any permission checks in userspace
based on ctx.uid/ctx.gid

[1] https://lore.kernel.org/linux-fsdevel/20240903151626.264609-1-aleksandr.mikhalitsyn@canonical.com
[2] https://kernel-recipes.org/en/2022/talks/idmapped-mounts/
[3] https://docs.kernel.org/next/filesystems/idmappings.html
[4] https://github.com/brauner/mount-idmapped
Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>