proc_paths: Use WrappedError type
We need this error type to implement a fall-back mechanism in
--migration-mode=find-paths: We want this mode to try to get inodes’
paths from /proc/self/fd first, and on certain kinds of errors to fall
back to iterating through the whole shared directory, collecting paths
that way.

For this, the /proc/self/fd code (proc_paths module) needs to be able to
distinguish between errors that are likely to appear regardless of how
we obtain inodes’ paths (i.e. falling back won’t fare better) and errors
that are likely tied to the specific method of getting paths from
/proc/self/fd, i.e. where falling back may succeed in finding a path.

Introduced the `WrappedError` type, and implement this distinction.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>