Move --migration-confirm-paths into own module
Create a new `device_state::preserialization::proc_paths` module that
implements the --migration-confirm-paths step: It iterates through all
inodes in our store, checks their migration infos’ paths, and when it
finds some error, tries to establish a new path from /proc/self/fd.

Note the name `proc_paths`: We are going to use code from this module
for the find-paths migration mode, to speed it up in most cases;
`proc_paths` because we get the *paths* from /*proc*/self/fd.

With this change, we no longer “confirm” the path precisely when an
inode is serialized, but in a separate loop before anything is
serialized.

Of note is that the code for finding a new path from /proc/self/fd is
rewritten: Specifically, before, we just wrote the path we got from
there directly into the migration stream, whereas now the new function
`set_path_migration_info_from_proc_self_fd()` splits the path into its
elements and updates all inodes’ migration info along the way (where
necessary).  That means a couple of things:
(1) Fixes a bug: The old code didn’t strip the shared directory prefix
    from the path, but probably should have.
(2) We no longer need the `serialized::InodeLocation::FullPath` variant.
    Leave it around for compatibility.
(3) We will be able to reuse this function for
    --migration-mode=find-paths.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>