Introduce VhostUserFsBackendBuilder
Introduce the VhostUserFsBackendBuilder type for customizable creation
of a VhostUserFsBackend instance. This makes it easier to change
otherwise sensible defaults on a per-attribute basis, while making it
easier to preserve backwards compatibility (adding a new argument to an
existing constructor would be a breaking change, but adding a new method
to the builder is not). Simplify the VhostUserFsBackend constructor with
sensible default values.

Signed-off-by: Daniel Müller <deso@posteo.net>