Introduce 'seccomp' feature
Cross-compilation of the crate and creation of statically linked
binaries based off of it are complicated by the fact that it links
against system libraries not written in Rust, seccomp & capng. In the
case of seccomp, it really should be left to clients to decide whether
the functionality is needed in the first place, at least when consuming
the library. It is already only contained in and exposed via a single
module, usage of which is entirely optional.

To give clients control over inclusion of seccomp functionality,
introduce the 'seccomp' feature, guarding availability of the module of
the same name. When disabled, libseccomp won't be linked.

Signed-off-by: Daniel Müller <deso@posteo.net>