Always clear O_NOATIME by default
Previously, if I ran virtiofsd as an unprivileged user, with the
sandbox enabled, and tried to open a file owned by root on the host in
the guest with O_NOATIME (for example by using overlayfs), virtiofsd
would respond with EPERM.  virtiofsd wouldn't clear O_NOATIME in this
case, because it had root inside the sandbox, but this only meant that
it could use O_NOATIME for files owned by the user running virtiofsd.

It's not really possible to detect when O_NOATIME will work without
trying it, so let's just drop the heuristic and have virtiofsd always
clear O_NOATIME unless told otherwise using the existing option.

Closes: https://gitlab.com/virtio-fs/virtiofsd/-/issues/180
Signed-off-by: Alyssa Ross <hi@alyssa.is>