Fix nlink type
Have the helper `link_count()` function return an `nlink_t` instead of
`u64`.  The actual representation of `nlink_t` is architecture-
dependent, so to return a `u64`, we may need to cast, which clippy likes
to warn about on those architectures where unnecessary.  That warning
would be silence-able, but it’s simpler to just return the actual
`nlink_t` type instead, without any cast.

Fixes: #183
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>