Use fuse::Attr in FileSystem interfaces
In the FileSystem interface, replace libc::stat64 by fuse::Attr.

With UID/GID mapping, converting one to the other is going to be
fallible, so server.rs can no longer do it (because generating guest
errors there is cumbersome and ugly), and we would rather have the
filesystem driver (PassthroughFs) attempt the conversion.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>