Make serialization infallible
Turning `PassthroughFs` into `serialized::PassthroughFsV2` is
effectively already infallible, albeit being implemented as `TryFrom`.

It really should be infallible, because errors should be communicated to
the migration destination instead of cancelling migration; change the
`TryFrom` to `From` to ensure it stays this way.

Note `<PassthroughFs as SerializableFileSystem>::serialize()` can still
fail in other places, but there is not much we can do about that:
- --migration-confirm-paths may fail to run altogether (shared directory
  inode missing from inode store (which would be an internal bug), or
  failing to find shared directory’s base path)
- Turning `serialized::PassthroughFsV2` into an actual byte vector (via
  serde/postcard) may fail, but that shouldn’t really happen
- Writing to the vhost-user migration data pipe may fail

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>