oslib::writev: Make outer doc comment an inner one
Today I’ve learned that `///` is called an “outer” comment, because it
is outside the described item, whereas `//!` is called an “inner”
comment because it is inside of the described item.

Anyway, clippy complains that this is an outer comment that currently
describes `pub use libc::pwritev2()`, but looks like it is supposed to
be an inner comment because of the empty space afterwards.  I agree, if
it is supposed to be an outer comment, it should be on the `pub fn
pwritev2()` for musl, if anything.

Making it an inner comment that describes the writev module and its
purpose makes sense, though, so do that.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>