filesystem: Remove unused and unsuitable methods
ZeroCopyReader::read_exact_to() and ::copy_to_end() are unused, same for
the respective methods in ZeroCopyWriter.  They are also unsuitable for
our needs:
- read_exact_to()/write_all_from() do not return the number actually
  read/written, instead returning an error if the underlying I/O
  operation returns 0.  We do want to allow partial I/O, though, but
  only if there is a good reason for it, i.e. because the underlying
  syscall returned 0 (end of file or no more space).
- copy_to_end() returns the value we want, but does not take a limiting
  length.

We are going to instead make write_from() and read_to() the methods we
need, which is not short reads/writes, unless there is a good reason for
it.

(We could change read_exact_to() and write_all_from() such that they
return the size actually written/read, but there is no reason to have
both this interface (the one that we want) and the short read/write
interface (which we don’t want) simultaneously, so it makes more sense
to scrap read_exact_to()/write_all_from() now and change
read_to()/write_from().)

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>