Rename write_from and read_to
This patch is based on a personal opinion, admittedly, but a strong one:
`read_to` and `write_from` are confusingly named.  Yes, they “read” data
from the descriptor chain into the file, or “write” data from the file
into the descriptor chain, but by that argument, you can call any read a
write and vice versa.

We call `read_to` in the FUSE WRITE operation, and what it does is write
to the file.  It should be called `write`.

We call `write_from` in the FUSE READ operation, and what it does is
read from the file.  It should be called `read`.

So rename `read_to` to `write_to_file_at`, and `write_from` to
`read_from_file_at`.  By including “file” in the name, the direction is
explicit and cannot be confusing.

Footnote: Yes, from a virtio device’s perspective, a read is writing
into memory, and vice versa.  However, I/O is generally given from the
guest driver’s perspective, and I believe we should follow that
convention.  See also virtio-blk, where IN is a read operation and OUT
is a write operation (read and write as defined by this patch).

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>