openat: Always annotate errors with filename
If the user ever sees the error string, it makes sense to have a
filename attached to it.

This could slightly reduce performance for failing guest-induced
open/lookup, in which case the error message is discarded anyway.  In
practice, I’m sure the whole operation (VM exit, syscall on our side)
required in the first place will much outweigh a `format!()` call.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>