Add and use openat_verbose() where reasonable
Instead of annotating all `openat()` calls with the filename, only do it
in a special new function `openat_verbose()`, which we are going to use
only where the error will be returned to the user and not to the guest,
i.e. where clobbering the raw OS error value doesn’t matter.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>