sandbox: fix pivot_root to relative paths
In the special case of shared_dir being a relative path (in my case
.), pivot_root(2) would fail in the previous implementation.  This
happened, as far as I can tell, because those paths would be resolved
from the process's CWD, which would still be referring to the original
mount, not the new one.  Absolute paths would resolve from the
process's root, so would end up pointing into the mount now at that
path.

Without doing an absolute path resolution, or leaving and re-entering
CWD with a .. path, both of which are not possible if the process does
not have permission to search the parent directory, I couldn't find
any way to get a file descriptor in the new mount created by mount(2).
I've solved the problem by using the new mount API, in which we first
create a detached mount, getting a file descriptor for it, and then
place it in the filesystem.  We can then pass that new file desciptor
to fchdir(), and pivot_root(2) will work.  The new mount APIs used
here are available since Linux 5.2, which is older than the Linux 5.3
already required for pidfd_open(2) by the namespace sandbox.

Note that it's not currently possible to reach this code with
self.shared_dir set to a non-absolute path, but that will change in
the next commit.

Signed-off-by: Alyssa Ross <hi@alyssa.is>