sandbox: only canonicalize shared_dir for chroot
According to d01a435 ("sandbox: Canonicalize shared_dir path"), which
introduced the fs::canonicalize() call here, canonicalization was only
known to be needed for the chroot sandbox mountinfo prefix, but there
weren't any known problems with doing it in all cases.

However, fs::canonicalize() will fail if any of the parent directories
of CWD are not searchable by the process, so for sandboxing modes
where we don't need it, avoid failing in that case by not calling
fs::canonicalize().

Signed-off-by: Alyssa Ross <hi@alyssa.is>