Update rust-vmm dependencies
Update rust-vmm dependencies, necessitating these changes:
- `VhostUserDaemon::start()` now takes a mutable reference to the
  `Listener` instead of ownership,
- The exit event `EventFd` has been replaced by an `EventConsumer` and
  `EventNotifier` pair.  We used to store `kill_evt` as an `EventFd` to
  write to in `VhostUserFsBackend::drop()`, but this should actually be
  already done by `VhostUserHandler::drop()` (in vhost-user-backend)
  ever since vhost commit f6b7e49e[1], so I think it’s time to drop it.
  `VhostUserHandler` directly owns our `VhostUserFsBackend` without
  further wrapping, so both are dropped at exactly the same time.

Note that this keeps vm-memory on 0.17.1 instead of 0.17.2 or 0.18.0: To
update to 0.18.0, we’d need some further changes to other rust-vmm
crates, and updating to 0.17.2 creates some conflicts in rust-vmm crates
because its re-export of `GuestAddressSpace` isn’t very useful (its
memory type is the new `GuestMemory`, but it would need to be the old
`GuestMemory` aka now `GuestMemoryBackend` to be useful).  I’ll send a
PR for the 0.17.2 issue to vm-memory, but for now just cap the version
at 0.17.1.

[1] https://github.com/rust-vmm/vhost/commit/f6b7e49e3f9ea
    (“handler: send exit event when destructing”)

Fixes: #96
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>